Data - Rename load_data to input_data to reflect class name.
